refactor: split paymentMethodService into validation, balance, and billing cycle sub-services
